requirements.txt
Pillow==9.3.0
requests==2.28.1
tqdm==4.64.1
